Change structure of folders
