+++ apps/block_scout_web/lib/block_scout_web/controllers/api/v2/optimism_controller.ex
@@ -9,7 +9,7 @@ defmodule BlockScoutWeb.API.V2.OptimismController do
     ]
 
   alias Explorer.Chain
-  alias Explorer.Chain.{OptimismDeposit, OptimismOutputRoot, OptimismTxnBatch, OptimismWithdrawal}
+  alias Explorer.Chain.Optimism.{Deposit, OutputRoot, TxnBatch, Withdrawal}
 
   action_fallback(BlockScoutWeb.API.V2.FallbackController)
 
@@ -32,7 +32,7 @@ defmodule BlockScoutWeb.API.V2.OptimismController do
   end
 
   def txn_batches_count(conn, _params) do
-    items_count(conn, OptimismTxnBatch)
+    items_count(conn, TxnBatch)
   end
 
   def output_roots(conn, params) do
@@ -54,7 +54,7 @@ defmodule BlockScoutWeb.API.V2.OptimismController do
   end
 
   def output_roots_count(conn, _params) do
-    items_count(conn, OptimismOutputRoot)
+    items_count(conn, OutputRoot)
   end
 
   def deposits(conn, params) do
@@ -76,7 +76,7 @@ defmodule BlockScoutWeb.API.V2.OptimismController do
   end
 
   def deposits_count(conn, _params) do
-    items_count(conn, OptimismDeposit)
+    items_count(conn, Deposit)
   end
 
   def withdrawals(conn, params) do
@@ -98,7 +98,7 @@ defmodule BlockScoutWeb.API.V2.OptimismController do
   end
 
   def withdrawals_count(conn, _params) do
-    items_count(conn, OptimismWithdrawal)
+    items_count(conn, Withdrawal)
   end
 
   defp items_count(conn, module) do
