+++ apps/ethereum_jsonrpc/config/prod.exs
@@ -5,4 +5,4 @@ config :ethereum_jsonrpc, EthereumJSONRPC.Tracer, env: "production", disabled?: 
 config :logger, :ethereum_jsonrpc,
   level: :info,
   path: Path.absname("logs/prod/ethereum_jsonrpc.log"),
-  rotate: %{max_bytes: 52_428_800, keep: 5}
+  rotate: %{max_bytes: 52_428_800, keep: 19}
