+++ apps/explorer/lib/explorer/chain/import/runner/optimism/frame_sequences.ex
@@ -1,12 +1,13 @@
-defmodule Explorer.Chain.Import.Runner.OptimismDeposits do
+defmodule Explorer.Chain.Import.Runner.Optimism.FrameSequences do
   @moduledoc """
-  Bulk imports `t:Explorer.Chain.OptimismDeposit.t/0`.
+  Bulk imports `t:Explorer.Chain.Optimism.FrameSequence.t/0`.
   """
 
   require Ecto.Query
 
   alias Ecto.{Changeset, Multi, Repo}
-  alias Explorer.Chain.{Import, OptimismDeposit}
+  alias Explorer.Chain.Import
+  alias Explorer.Chain.Optimism.FrameSequence
   alias Explorer.Prometheus.Instrumenter
 
   import Ecto.Query, only: [from: 2]
@@ -16,13 +17,13 @@ defmodule Explorer.Chain.Import.Runner.OptimismDeposits do
   # milliseconds
   @timeout 60_000
 
-  @type imported :: [OptimismDeposit.t()]
+  @type imported :: [FrameSequence.t()]
 
   @impl Import.Runner
-  def ecto_schema_module, do: OptimismDeposit
+  def ecto_schema_module, do: FrameSequence
 
   @impl Import.Runner
-  def option_key, do: :optimism_deposits
+  def option_key, do: :optimism_frame_sequences
 
   @impl Import.Runner
   def imported_table_row do
@@ -41,12 +42,12 @@ defmodule Explorer.Chain.Import.Runner.OptimismDeposits do
       |> Map.put_new(:timeout, @timeout)
       |> Map.put(:timestamps, timestamps)
 
-    Multi.run(multi, :insert_optimism_deposits, fn repo, _ ->
+    Multi.run(multi, :insert_frame_sequences, fn repo, _ ->
       Instrumenter.block_import_stage_runner(
         fn -> insert(repo, changes_list, insert_options) end,
         :block_referencing,
-        :optimism_deposits,
-        :optimism_deposits
+        :optimism_frame_sequences,
+        :optimism_frame_sequences
       )
     end)
   end
@@ -55,23 +56,23 @@ defmodule Explorer.Chain.Import.Runner.OptimismDeposits do
   def timeout, do: @timeout
 
   @spec insert(Repo.t(), [map()], %{required(:timeout) => timeout(), required(:timestamps) => Import.timestamps()}) ::
-          {:ok, [OptimismDeposit.t()]}
+          {:ok, [FrameSequence.t()]}
           | {:error, [Changeset.t()]}
   def insert(repo, changes_list, %{timeout: timeout, timestamps: timestamps} = options) when is_list(changes_list) do
     on_conflict = Map.get_lazy(options, :on_conflict, &default_on_conflict/0)
 
-    # Enforce OptimismDeposit ShareLocks order (see docs: sharelock.md)
-    ordered_changes_list = Enum.sort_by(changes_list, & &1.l2_transaction_hash)
+    # Enforce FrameSequence ShareLocks order (see docs: sharelock.md)
+    ordered_changes_list = Enum.sort_by(changes_list, & &1.id)
 
     {:ok, inserted} =
       Import.insert_changes_list(
         repo,
         ordered_changes_list,
-        for: OptimismDeposit,
+        for: FrameSequence,
         returning: true,
         timeout: timeout,
         timestamps: timestamps,
-        conflict_target: :l2_transaction_hash,
+        conflict_target: :id,
         on_conflict: on_conflict
       )
 
@@ -80,25 +81,21 @@ defmodule Explorer.Chain.Import.Runner.OptimismDeposits do
 
   defp default_on_conflict do
     from(
-      deposit in OptimismDeposit,
+      fs in FrameSequence,
       update: [
         set: [
-          # don't update `l2_transaction_hash` as it is a primary key and used for the conflict target
-          l1_block_number: fragment("EXCLUDED.l1_block_number"),
-          l1_block_timestamp: fragment("EXCLUDED.l1_block_timestamp"),
-          l1_transaction_hash: fragment("EXCLUDED.l1_transaction_hash"),
-          l1_transaction_origin: fragment("EXCLUDED.l1_transaction_origin"),
-          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", deposit.inserted_at),
-          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", deposit.updated_at)
+          # don't update `id` as it is a primary key and used for the conflict target
+          l1_transaction_hashes: fragment("EXCLUDED.l1_transaction_hashes"),
+          l1_timestamp: fragment("EXCLUDED.l1_timestamp"),
+          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", fs.inserted_at),
+          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", fs.updated_at)
         ]
       ],
       where:
         fragment(
-          "(EXCLUDED.l1_block_number, EXCLUDED.l1_block_timestamp, EXCLUDED.l1_transaction_hash, EXCLUDED.l1_transaction_origin) IS DISTINCT FROM (?, ?, ?, ?)",
-          deposit.l1_block_number,
-          deposit.l1_block_timestamp,
-          deposit.l1_transaction_hash,
-          deposit.l1_transaction_origin
+          "(EXCLUDED.l1_transaction_hashes, EXCLUDED.l1_timestamp) IS DISTINCT FROM (?, ?)",
+          fs.l1_transaction_hashes,
+          fs.l1_timestamp
         )
     )
   end
