+++ apps/explorer/lib/explorer/chain/import/runner/optimism/txn_batches.ex
@@ -1,12 +1,13 @@
-defmodule Explorer.Chain.Import.Runner.OptimismWithdrawalEvents do
+defmodule Explorer.Chain.Import.Runner.Optimism.TxnBatches do
   @moduledoc """
-  Bulk imports `t:Explorer.Chain.OptimismWithdrawalEvent.t/0`.
+  Bulk imports `t:Explorer.Chain.Optimism.TxnBatch.t/0`.
   """
 
   require Ecto.Query
 
   alias Ecto.{Changeset, Multi, Repo}
-  alias Explorer.Chain.{Import, OptimismWithdrawalEvent}
+  alias Explorer.Chain.Import
+  alias Explorer.Chain.Optimism.TxnBatch
   alias Explorer.Prometheus.Instrumenter
 
   import Ecto.Query, only: [from: 2]
@@ -16,13 +17,13 @@ defmodule Explorer.Chain.Import.Runner.OptimismWithdrawalEvents do
   # milliseconds
   @timeout 60_000
 
-  @type imported :: [OptimismWithdrawalEvent.t()]
+  @type imported :: [TxnBatch.t()]
 
   @impl Import.Runner
-  def ecto_schema_module, do: OptimismWithdrawalEvent
+  def ecto_schema_module, do: TxnBatch
 
   @impl Import.Runner
-  def option_key, do: :optimism_withdrawal_events
+  def option_key, do: :optimism_txn_batches
 
   @impl Import.Runner
   def imported_table_row do
@@ -41,12 +42,12 @@ defmodule Explorer.Chain.Import.Runner.OptimismWithdrawalEvents do
       |> Map.put_new(:timeout, @timeout)
       |> Map.put(:timestamps, timestamps)
 
-    Multi.run(multi, :insert_withdrawal_events, fn repo, _ ->
+    Multi.run(multi, :insert_txn_batches, fn repo, _ ->
       Instrumenter.block_import_stage_runner(
         fn -> insert(repo, changes_list, insert_options) end,
         :block_referencing,
-        :optimism_withdrawal_events,
-        :optimism_withdrawal_events
+        :optimism_txn_batches,
+        :optimism_txn_batches
       )
     end)
   end
@@ -55,23 +56,23 @@ defmodule Explorer.Chain.Import.Runner.OptimismWithdrawalEvents do
   def timeout, do: @timeout
 
   @spec insert(Repo.t(), [map()], %{required(:timeout) => timeout(), required(:timestamps) => Import.timestamps()}) ::
-          {:ok, [OptimismWithdrawalEvent.t()]}
+          {:ok, [TxnBatch.t()]}
           | {:error, [Changeset.t()]}
   def insert(repo, changes_list, %{timeout: timeout, timestamps: timestamps} = options) when is_list(changes_list) do
     on_conflict = Map.get_lazy(options, :on_conflict, &default_on_conflict/0)
 
-    # Enforce OptimismWithdrawalEvent ShareLocks order (see docs: sharelock.md)
-    ordered_changes_list = Enum.sort_by(changes_list, &{&1.withdrawal_hash, &1.l1_event_type})
+    # Enforce TxnBatch ShareLocks order (see docs: sharelock.md)
+    ordered_changes_list = Enum.sort_by(changes_list, & &1.l2_block_number)
 
     {:ok, inserted} =
       Import.insert_changes_list(
         repo,
         ordered_changes_list,
-        for: OptimismWithdrawalEvent,
+        for: TxnBatch,
         returning: true,
         timeout: timeout,
         timestamps: timestamps,
-        conflict_target: [:withdrawal_hash, :l1_event_type],
+        conflict_target: :l2_block_number,
         on_conflict: on_conflict
       )
 
@@ -80,24 +81,19 @@ defmodule Explorer.Chain.Import.Runner.OptimismWithdrawalEvents do
 
   defp default_on_conflict do
     from(
-      we in OptimismWithdrawalEvent,
+      tb in TxnBatch,
       update: [
         set: [
-          # don't update `withdrawal_hash` as it is a part of the composite primary key and used for the conflict target
-          # don't update `l1_event_type` as it is a part of the composite primary key and used for the conflict target
-          l1_timestamp: fragment("EXCLUDED.l1_timestamp"),
-          l1_transaction_hash: fragment("EXCLUDED.l1_transaction_hash"),
-          l1_block_number: fragment("EXCLUDED.l1_block_number"),
-          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", we.inserted_at),
-          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", we.updated_at)
+          # don't update `l2_block_number` as it is a primary key and used for the conflict target
+          frame_sequence_id: fragment("EXCLUDED.frame_sequence_id"),
+          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", tb.inserted_at),
+          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", tb.updated_at)
         ]
       ],
       where:
         fragment(
-          "(EXCLUDED.l1_timestamp, EXCLUDED.l1_transaction_hash, EXCLUDED.l1_block_number) IS DISTINCT FROM (?, ?, ?)",
-          we.l1_timestamp,
-          we.l1_transaction_hash,
-          we.l1_block_number
+          "(EXCLUDED.frame_sequence_id) IS DISTINCT FROM (?)",
+          tb.frame_sequence_id
         )
     )
   end
