+++ apps/explorer/lib/explorer/chain/import/runner/optimism/withdrawals.ex
@@ -1,12 +1,13 @@
-defmodule Explorer.Chain.Import.Runner.OptimismTxnBatches do
+defmodule Explorer.Chain.Import.Runner.Optimism.Withdrawals do
   @moduledoc """
-  Bulk imports `t:Explorer.Chain.OptimismTxnBatch.t/0`.
+  Bulk imports `t:Explorer.Chain.OptimismWithdrawal.t/0`.
   """
 
   require Ecto.Query
 
   alias Ecto.{Changeset, Multi, Repo}
-  alias Explorer.Chain.{Import, OptimismTxnBatch}
+  alias Explorer.Chain.Import
+  alias Explorer.Chain.Optimism.Withdrawal, as: OptimismWithdrawal
   alias Explorer.Prometheus.Instrumenter
 
   import Ecto.Query, only: [from: 2]
@@ -16,13 +17,13 @@ defmodule Explorer.Chain.Import.Runner.OptimismTxnBatches do
   # milliseconds
   @timeout 60_000
 
-  @type imported :: [OptimismTxnBatch.t()]
+  @type imported :: [OptimismWithdrawal.t()]
 
   @impl Import.Runner
-  def ecto_schema_module, do: OptimismTxnBatch
+  def ecto_schema_module, do: OptimismWithdrawal
 
   @impl Import.Runner
-  def option_key, do: :optimism_txn_batches
+  def option_key, do: :optimism_withdrawals
 
   @impl Import.Runner
   def imported_table_row do
@@ -41,12 +42,12 @@ defmodule Explorer.Chain.Import.Runner.OptimismTxnBatches do
       |> Map.put_new(:timeout, @timeout)
       |> Map.put(:timestamps, timestamps)
 
-    Multi.run(multi, :insert_txn_batches, fn repo, _ ->
+    Multi.run(multi, :insert_withdrawals, fn repo, _ ->
       Instrumenter.block_import_stage_runner(
         fn -> insert(repo, changes_list, insert_options) end,
         :block_referencing,
-        :optimism_txn_batches,
-        :optimism_txn_batches
+        :optimism_withdrawals,
+        :optimism_withdrawals
       )
     end)
   end
@@ -55,23 +56,23 @@ defmodule Explorer.Chain.Import.Runner.OptimismTxnBatches do
   def timeout, do: @timeout
 
   @spec insert(Repo.t(), [map()], %{required(:timeout) => timeout(), required(:timestamps) => Import.timestamps()}) ::
-          {:ok, [OptimismTxnBatch.t()]}
+          {:ok, [OptimismWithdrawal.t()]}
           | {:error, [Changeset.t()]}
   def insert(repo, changes_list, %{timeout: timeout, timestamps: timestamps} = options) when is_list(changes_list) do
     on_conflict = Map.get_lazy(options, :on_conflict, &default_on_conflict/0)
 
-    # Enforce OptimismTxnBatch ShareLocks order (see docs: sharelock.md)
-    ordered_changes_list = Enum.sort_by(changes_list, & &1.l2_block_number)
+    # Enforce OptimismWithdrawal ShareLocks order (see docs: sharelock.md)
+    ordered_changes_list = Enum.sort_by(changes_list, & &1.msg_nonce)
 
     {:ok, inserted} =
       Import.insert_changes_list(
         repo,
         ordered_changes_list,
-        for: OptimismTxnBatch,
+        for: OptimismWithdrawal,
         returning: true,
         timeout: timeout,
         timestamps: timestamps,
-        conflict_target: :l2_block_number,
+        conflict_target: :msg_nonce,
         on_conflict: on_conflict
       )
 
@@ -80,19 +81,23 @@ defmodule Explorer.Chain.Import.Runner.OptimismTxnBatches do
 
   defp default_on_conflict do
     from(
-      tb in OptimismTxnBatch,
+      withdrawal in OptimismWithdrawal,
       update: [
         set: [
-          # don't update `l2_block_number` as it is a primary key and used for the conflict target
-          frame_sequence_id: fragment("EXCLUDED.frame_sequence_id"),
-          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", tb.inserted_at),
-          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", tb.updated_at)
+          # don't update `msg_nonce` as it is a primary key and used for the conflict target
+          hash: fragment("EXCLUDED.hash"),
+          l2_transaction_hash: fragment("EXCLUDED.l2_transaction_hash"),
+          l2_block_number: fragment("EXCLUDED.l2_block_number"),
+          inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", withdrawal.inserted_at),
+          updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", withdrawal.updated_at)
         ]
       ],
       where:
         fragment(
-          "(EXCLUDED.frame_sequence_id) IS DISTINCT FROM (?)",
-          tb.frame_sequence_id
+          "(EXCLUDED.hash, EXCLUDED.l2_transaction_hash, EXCLUDED.l2_block_number) IS DISTINCT FROM (?, ?, ?)",
+          withdrawal.hash,
+          withdrawal.l2_transaction_hash,
+          withdrawal.l2_block_number
         )
     )
   end
