+++ apps/explorer/lib/explorer/chain/import/runner/transactions.ex
@@ -108,192 +108,250 @@ defmodule Explorer.Chain.Import.Runner.Transactions do
   end
 
   defp default_on_conflict do
-    if System.get_env("CHAIN_TYPE") == "suave" do
-      from(
-        transaction in Transaction,
-        update: [
-          set: [
-            block_hash: fragment("EXCLUDED.block_hash"),
-            old_block_hash: transaction.block_hash,
-            block_number: fragment("EXCLUDED.block_number"),
-            block_consensus: fragment("EXCLUDED.block_consensus"),
-            block_timestamp: fragment("EXCLUDED.block_timestamp"),
-            created_contract_address_hash: fragment("EXCLUDED.created_contract_address_hash"),
-            created_contract_code_indexed_at: fragment("EXCLUDED.created_contract_code_indexed_at"),
-            cumulative_gas_used: fragment("EXCLUDED.cumulative_gas_used"),
-            error: fragment("EXCLUDED.error"),
-            from_address_hash: fragment("EXCLUDED.from_address_hash"),
-            gas: fragment("EXCLUDED.gas"),
-            gas_price: fragment("EXCLUDED.gas_price"),
-            gas_used: fragment("EXCLUDED.gas_used"),
-            index: fragment("EXCLUDED.index"),
-            input: fragment("EXCLUDED.input"),
-            nonce: fragment("EXCLUDED.nonce"),
-            r: fragment("EXCLUDED.r"),
-            s: fragment("EXCLUDED.s"),
-            status: fragment("EXCLUDED.status"),
-            to_address_hash: fragment("EXCLUDED.to_address_hash"),
-            v: fragment("EXCLUDED.v"),
-            value: fragment("EXCLUDED.value"),
-            earliest_processing_start: fragment("EXCLUDED.earliest_processing_start"),
-            revert_reason: fragment("EXCLUDED.revert_reason"),
-            max_priority_fee_per_gas: fragment("EXCLUDED.max_priority_fee_per_gas"),
-            max_fee_per_gas: fragment("EXCLUDED.max_fee_per_gas"),
-            type: fragment("EXCLUDED.type"),
-            execution_node_hash: fragment("EXCLUDED.execution_node_hash"),
-            wrapped_type: fragment("EXCLUDED.wrapped_type"),
-            wrapped_nonce: fragment("EXCLUDED.wrapped_nonce"),
-            wrapped_to_address_hash: fragment("EXCLUDED.wrapped_to_address_hash"),
-            wrapped_gas: fragment("EXCLUDED.wrapped_gas"),
-            wrapped_gas_price: fragment("EXCLUDED.wrapped_gas_price"),
-            wrapped_max_priority_fee_per_gas: fragment("EXCLUDED.wrapped_max_priority_fee_per_gas"),
-            wrapped_max_fee_per_gas: fragment("EXCLUDED.wrapped_max_fee_per_gas"),
-            wrapped_value: fragment("EXCLUDED.wrapped_value"),
-            wrapped_input: fragment("EXCLUDED.wrapped_input"),
-            wrapped_v: fragment("EXCLUDED.wrapped_v"),
-            wrapped_r: fragment("EXCLUDED.wrapped_r"),
-            wrapped_s: fragment("EXCLUDED.wrapped_s"),
-            wrapped_hash: fragment("EXCLUDED.wrapped_hash"),
-            l1_fee: fragment("EXCLUDED.l1_fee"),
-            l1_fee_scalar: fragment("EXCLUDED.l1_fee_scalar"),
-            l1_gas_price: fragment("EXCLUDED.l1_gas_price"),
-            l1_gas_used: fragment("EXCLUDED.l1_gas_used"),
-            l1_tx_origin: fragment("EXCLUDED.l1_tx_origin"),
-            l1_block_number: fragment("EXCLUDED.l1_block_number"),
-            # Don't update `hash` as it is part of the primary key and used for the conflict target
-            inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", transaction.inserted_at),
-            updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", transaction.updated_at)
-          ]
-        ],
-        where:
-          fragment(
-            "(EXCLUDED.block_hash, EXCLUDED.block_number, EXCLUDED.block_consensus, EXCLUDED.block_timestamp, EXCLUDED.created_contract_address_hash, EXCLUDED.created_contract_code_indexed_at, EXCLUDED.cumulative_gas_used, EXCLUDED.from_address_hash, EXCLUDED.gas, EXCLUDED.gas_price, EXCLUDED.gas_used, EXCLUDED.index, EXCLUDED.input, EXCLUDED.nonce, EXCLUDED.r, EXCLUDED.s, EXCLUDED.status, EXCLUDED.to_address_hash, EXCLUDED.v, EXCLUDED.value, EXCLUDED.earliest_processing_start, EXCLUDED.revert_reason, EXCLUDED.max_priority_fee_per_gas, EXCLUDED.max_fee_per_gas, EXCLUDED.type, EXCLUDED.execution_node_hash, EXCLUDED.wrapped_type, EXCLUDED.wrapped_nonce, EXCLUDED.wrapped_to_address_hash, EXCLUDED.wrapped_gas, EXCLUDED.wrapped_gas_price, EXCLUDED.wrapped_max_priority_fee_per_gas, EXCLUDED.wrapped_max_fee_per_gas, EXCLUDED.wrapped_value, EXCLUDED.wrapped_input, EXCLUDED.wrapped_v, EXCLUDED.wrapped_r, EXCLUDED.wrapped_s, EXCLUDED.wrapped_hash, EXCLUDED.l1_fee, EXCLUDED.l1_fee_scalar, EXCLUDED.l1_gas_price, EXCLUDED.l1_gas_used, EXCLUDED.l1_tx_origin, EXCLUDED.l1_block_number) IS DISTINCT FROM (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-            transaction.block_hash,
-            transaction.block_number,
-            transaction.block_consensus,
-            transaction.block_timestamp,
-            transaction.created_contract_address_hash,
-            transaction.created_contract_code_indexed_at,
-            transaction.cumulative_gas_used,
-            transaction.from_address_hash,
-            transaction.gas,
-            transaction.gas_price,
-            transaction.gas_used,
-            transaction.index,
-            transaction.input,
-            transaction.nonce,
-            transaction.r,
-            transaction.s,
-            transaction.status,
-            transaction.to_address_hash,
-            transaction.v,
-            transaction.value,
-            transaction.earliest_processing_start,
-            transaction.revert_reason,
-            transaction.max_priority_fee_per_gas,
-            transaction.max_fee_per_gas,
-            transaction.type,
-            transaction.execution_node_hash,
-            transaction.wrapped_type,
-            transaction.wrapped_nonce,
-            transaction.wrapped_to_address_hash,
-            transaction.wrapped_gas,
-            transaction.wrapped_gas_price,
-            transaction.wrapped_max_priority_fee_per_gas,
-            transaction.wrapped_max_fee_per_gas,
-            transaction.wrapped_value,
-            transaction.wrapped_input,
-            transaction.wrapped_v,
-            transaction.wrapped_r,
-            transaction.wrapped_s,
-            transaction.wrapped_hash,
-            transaction.l1_fee,
-            transaction.l1_fee_scalar,
-            transaction.l1_gas_price,
-            transaction.l1_gas_used,
-            transaction.l1_tx_origin,
-            transaction.l1_block_number
-          )
-      )
-    else
-      from(
-        transaction in Transaction,
-        update: [
-          set: [
-            block_hash: fragment("EXCLUDED.block_hash"),
-            old_block_hash: transaction.block_hash,
-            block_number: fragment("EXCLUDED.block_number"),
-            block_consensus: fragment("EXCLUDED.block_consensus"),
-            block_timestamp: fragment("EXCLUDED.block_timestamp"),
-            created_contract_address_hash: fragment("EXCLUDED.created_contract_address_hash"),
-            created_contract_code_indexed_at: fragment("EXCLUDED.created_contract_code_indexed_at"),
-            cumulative_gas_used: fragment("EXCLUDED.cumulative_gas_used"),
-            error: fragment("EXCLUDED.error"),
-            from_address_hash: fragment("EXCLUDED.from_address_hash"),
-            gas: fragment("EXCLUDED.gas"),
-            gas_price: fragment("EXCLUDED.gas_price"),
-            gas_used: fragment("EXCLUDED.gas_used"),
-            index: fragment("EXCLUDED.index"),
-            input: fragment("EXCLUDED.input"),
-            nonce: fragment("EXCLUDED.nonce"),
-            r: fragment("EXCLUDED.r"),
-            s: fragment("EXCLUDED.s"),
-            status: fragment("EXCLUDED.status"),
-            to_address_hash: fragment("EXCLUDED.to_address_hash"),
-            v: fragment("EXCLUDED.v"),
-            value: fragment("EXCLUDED.value"),
-            earliest_processing_start: fragment("EXCLUDED.earliest_processing_start"),
-            revert_reason: fragment("EXCLUDED.revert_reason"),
-            max_priority_fee_per_gas: fragment("EXCLUDED.max_priority_fee_per_gas"),
-            max_fee_per_gas: fragment("EXCLUDED.max_fee_per_gas"),
-            type: fragment("EXCLUDED.type"),
-            l1_fee: fragment("EXCLUDED.l1_fee"),
-            l1_fee_scalar: fragment("EXCLUDED.l1_fee_scalar"),
-            l1_gas_price: fragment("EXCLUDED.l1_gas_price"),
-            l1_gas_used: fragment("EXCLUDED.l1_gas_used"),
-            l1_tx_origin: fragment("EXCLUDED.l1_tx_origin"),
-            l1_block_number: fragment("EXCLUDED.l1_block_number"),
-            # Don't update `hash` as it is part of the primary key and used for the conflict target
-            inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", transaction.inserted_at),
-            updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", transaction.updated_at)
-          ]
-        ],
-        where:
-          fragment(
-            "(EXCLUDED.block_hash, EXCLUDED.block_number, EXCLUDED.block_consensus, EXCLUDED.block_timestamp, EXCLUDED.created_contract_address_hash, EXCLUDED.created_contract_code_indexed_at, EXCLUDED.cumulative_gas_used, EXCLUDED.from_address_hash, EXCLUDED.gas, EXCLUDED.gas_price, EXCLUDED.gas_used, EXCLUDED.index, EXCLUDED.input, EXCLUDED.nonce, EXCLUDED.r, EXCLUDED.s, EXCLUDED.status, EXCLUDED.to_address_hash, EXCLUDED.v, EXCLUDED.value, EXCLUDED.earliest_processing_start, EXCLUDED.revert_reason, EXCLUDED.max_priority_fee_per_gas, EXCLUDED.max_fee_per_gas, EXCLUDED.type, EXCLUDED.l1_fee, EXCLUDED.l1_fee_scalar, EXCLUDED.l1_gas_price, EXCLUDED.l1_gas_used, EXCLUDED.l1_tx_origin, EXCLUDED.l1_block_number) IS DISTINCT FROM (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
-            transaction.block_hash,
-            transaction.block_number,
-            transaction.block_consensus,
-            transaction.block_timestamp,
-            transaction.created_contract_address_hash,
-            transaction.created_contract_code_indexed_at,
-            transaction.cumulative_gas_used,
-            transaction.from_address_hash,
-            transaction.gas,
-            transaction.gas_price,
-            transaction.gas_used,
-            transaction.index,
-            transaction.input,
-            transaction.nonce,
-            transaction.r,
-            transaction.s,
-            transaction.status,
-            transaction.to_address_hash,
-            transaction.v,
-            transaction.value,
-            transaction.earliest_processing_start,
-            transaction.revert_reason,
-            transaction.max_priority_fee_per_gas,
-            transaction.max_fee_per_gas,
-            transaction.type,
-            transaction.l1_fee,
-            transaction.l1_fee_scalar,
-            transaction.l1_gas_price,
-            transaction.l1_gas_used,
-            transaction.l1_tx_origin,
-            transaction.l1_block_number
-          )
-      )
+    case System.get_env("CHAIN_TYPE") do
+      "suave" ->
+        from(
+          transaction in Transaction,
+          update: [
+            set: [
+              block_hash: fragment("EXCLUDED.block_hash"),
+              old_block_hash: transaction.block_hash,
+              block_number: fragment("EXCLUDED.block_number"),
+              block_consensus: fragment("EXCLUDED.block_consensus"),
+              block_timestamp: fragment("EXCLUDED.block_timestamp"),
+              created_contract_address_hash: fragment("EXCLUDED.created_contract_address_hash"),
+              created_contract_code_indexed_at: fragment("EXCLUDED.created_contract_code_indexed_at"),
+              cumulative_gas_used: fragment("EXCLUDED.cumulative_gas_used"),
+              error: fragment("EXCLUDED.error"),
+              from_address_hash: fragment("EXCLUDED.from_address_hash"),
+              gas: fragment("EXCLUDED.gas"),
+              gas_price: fragment("EXCLUDED.gas_price"),
+              gas_used: fragment("EXCLUDED.gas_used"),
+              index: fragment("EXCLUDED.index"),
+              input: fragment("EXCLUDED.input"),
+              nonce: fragment("EXCLUDED.nonce"),
+              r: fragment("EXCLUDED.r"),
+              s: fragment("EXCLUDED.s"),
+              status: fragment("EXCLUDED.status"),
+              to_address_hash: fragment("EXCLUDED.to_address_hash"),
+              v: fragment("EXCLUDED.v"),
+              value: fragment("EXCLUDED.value"),
+              earliest_processing_start: fragment("EXCLUDED.earliest_processing_start"),
+              revert_reason: fragment("EXCLUDED.revert_reason"),
+              max_priority_fee_per_gas: fragment("EXCLUDED.max_priority_fee_per_gas"),
+              max_fee_per_gas: fragment("EXCLUDED.max_fee_per_gas"),
+              type: fragment("EXCLUDED.type"),
+              execution_node_hash: fragment("EXCLUDED.execution_node_hash"),
+              wrapped_type: fragment("EXCLUDED.wrapped_type"),
+              wrapped_nonce: fragment("EXCLUDED.wrapped_nonce"),
+              wrapped_to_address_hash: fragment("EXCLUDED.wrapped_to_address_hash"),
+              wrapped_gas: fragment("EXCLUDED.wrapped_gas"),
+              wrapped_gas_price: fragment("EXCLUDED.wrapped_gas_price"),
+              wrapped_max_priority_fee_per_gas: fragment("EXCLUDED.wrapped_max_priority_fee_per_gas"),
+              wrapped_max_fee_per_gas: fragment("EXCLUDED.wrapped_max_fee_per_gas"),
+              wrapped_value: fragment("EXCLUDED.wrapped_value"),
+              wrapped_input: fragment("EXCLUDED.wrapped_input"),
+              wrapped_v: fragment("EXCLUDED.wrapped_v"),
+              wrapped_r: fragment("EXCLUDED.wrapped_r"),
+              wrapped_s: fragment("EXCLUDED.wrapped_s"),
+              wrapped_hash: fragment("EXCLUDED.wrapped_hash"),
+              # Don't update `hash` as it is part of the primary key and used for the conflict target
+              inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", transaction.inserted_at),
+              updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", transaction.updated_at)
+            ]
+          ],
+          where:
+            fragment(
+              "(EXCLUDED.block_hash, EXCLUDED.block_number, EXCLUDED.block_consensus, EXCLUDED.block_timestamp, EXCLUDED.created_contract_address_hash, EXCLUDED.created_contract_code_indexed_at, EXCLUDED.cumulative_gas_used, EXCLUDED.from_address_hash, EXCLUDED.gas, EXCLUDED.gas_price, EXCLUDED.gas_used, EXCLUDED.index, EXCLUDED.input, EXCLUDED.nonce, EXCLUDED.r, EXCLUDED.s, EXCLUDED.status, EXCLUDED.to_address_hash, EXCLUDED.v, EXCLUDED.value, EXCLUDED.earliest_processing_start, EXCLUDED.revert_reason, EXCLUDED.max_priority_fee_per_gas, EXCLUDED.max_fee_per_gas, EXCLUDED.type, EXCLUDED.execution_node_hash, EXCLUDED.wrapped_type, EXCLUDED.wrapped_nonce, EXCLUDED.wrapped_to_address_hash, EXCLUDED.wrapped_gas, EXCLUDED.wrapped_gas_price, EXCLUDED.wrapped_max_priority_fee_per_gas, EXCLUDED.wrapped_max_fee_per_gas, EXCLUDED.wrapped_value, EXCLUDED.wrapped_input, EXCLUDED.wrapped_v, EXCLUDED.wrapped_r, EXCLUDED.wrapped_s, EXCLUDED.wrapped_hash) IS DISTINCT FROM (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+              transaction.block_hash,
+              transaction.block_number,
+              transaction.block_consensus,
+              transaction.block_timestamp,
+              transaction.created_contract_address_hash,
+              transaction.created_contract_code_indexed_at,
+              transaction.cumulative_gas_used,
+              transaction.from_address_hash,
+              transaction.gas,
+              transaction.gas_price,
+              transaction.gas_used,
+              transaction.index,
+              transaction.input,
+              transaction.nonce,
+              transaction.r,
+              transaction.s,
+              transaction.status,
+              transaction.to_address_hash,
+              transaction.v,
+              transaction.value,
+              transaction.earliest_processing_start,
+              transaction.revert_reason,
+              transaction.max_priority_fee_per_gas,
+              transaction.max_fee_per_gas,
+              transaction.type,
+              transaction.execution_node_hash,
+              transaction.wrapped_type,
+              transaction.wrapped_nonce,
+              transaction.wrapped_to_address_hash,
+              transaction.wrapped_gas,
+              transaction.wrapped_gas_price,
+              transaction.wrapped_max_priority_fee_per_gas,
+              transaction.wrapped_max_fee_per_gas,
+              transaction.wrapped_value,
+              transaction.wrapped_input,
+              transaction.wrapped_v,
+              transaction.wrapped_r,
+              transaction.wrapped_s,
+              transaction.wrapped_hash
+            )
+        )
+
+      "optimism" ->
+        from(
+          transaction in Transaction,
+          update: [
+            set: [
+              block_hash: fragment("EXCLUDED.block_hash"),
+              old_block_hash: transaction.block_hash,
+              block_number: fragment("EXCLUDED.block_number"),
+              block_consensus: fragment("EXCLUDED.block_consensus"),
+              block_timestamp: fragment("EXCLUDED.block_timestamp"),
+              created_contract_address_hash: fragment("EXCLUDED.created_contract_address_hash"),
+              created_contract_code_indexed_at: fragment("EXCLUDED.created_contract_code_indexed_at"),
+              cumulative_gas_used: fragment("EXCLUDED.cumulative_gas_used"),
+              error: fragment("EXCLUDED.error"),
+              from_address_hash: fragment("EXCLUDED.from_address_hash"),
+              gas: fragment("EXCLUDED.gas"),
+              gas_price: fragment("EXCLUDED.gas_price"),
+              gas_used: fragment("EXCLUDED.gas_used"),
+              index: fragment("EXCLUDED.index"),
+              input: fragment("EXCLUDED.input"),
+              nonce: fragment("EXCLUDED.nonce"),
+              r: fragment("EXCLUDED.r"),
+              s: fragment("EXCLUDED.s"),
+              status: fragment("EXCLUDED.status"),
+              to_address_hash: fragment("EXCLUDED.to_address_hash"),
+              v: fragment("EXCLUDED.v"),
+              value: fragment("EXCLUDED.value"),
+              earliest_processing_start: fragment("EXCLUDED.earliest_processing_start"),
+              revert_reason: fragment("EXCLUDED.revert_reason"),
+              max_priority_fee_per_gas: fragment("EXCLUDED.max_priority_fee_per_gas"),
+              max_fee_per_gas: fragment("EXCLUDED.max_fee_per_gas"),
+              type: fragment("EXCLUDED.type"),
+              l1_fee: fragment("EXCLUDED.l1_fee"),
+              l1_fee_scalar: fragment("EXCLUDED.l1_fee_scalar"),
+              l1_gas_price: fragment("EXCLUDED.l1_gas_price"),
+              l1_gas_used: fragment("EXCLUDED.l1_gas_used"),
+              l1_tx_origin: fragment("EXCLUDED.l1_tx_origin"),
+              l1_block_number: fragment("EXCLUDED.l1_block_number"),
+              # Don't update `hash` as it is part of the primary key and used for the conflict target
+              inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", transaction.inserted_at),
+              updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", transaction.updated_at)
+            ]
+          ],
+          where:
+            fragment(
+              "(EXCLUDED.block_hash, EXCLUDED.block_number, EXCLUDED.block_consensus, EXCLUDED.block_timestamp, EXCLUDED.created_contract_address_hash, EXCLUDED.created_contract_code_indexed_at, EXCLUDED.cumulative_gas_used, EXCLUDED.from_address_hash, EXCLUDED.gas, EXCLUDED.gas_price, EXCLUDED.gas_used, EXCLUDED.index, EXCLUDED.input, EXCLUDED.nonce, EXCLUDED.r, EXCLUDED.s, EXCLUDED.status, EXCLUDED.to_address_hash, EXCLUDED.v, EXCLUDED.value, EXCLUDED.earliest_processing_start, EXCLUDED.revert_reason, EXCLUDED.max_priority_fee_per_gas, EXCLUDED.max_fee_per_gas, EXCLUDED.type, EXCLUDED.l1_fee, EXCLUDED.l1_fee_scalar, EXCLUDED.l1_gas_price, EXCLUDED.l1_gas_used, EXCLUDED.l1_tx_origin, EXCLUDED.l1_block_number) IS DISTINCT FROM (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+              transaction.block_hash,
+              transaction.block_number,
+              transaction.block_consensus,
+              transaction.block_timestamp,
+              transaction.created_contract_address_hash,
+              transaction.created_contract_code_indexed_at,
+              transaction.cumulative_gas_used,
+              transaction.from_address_hash,
+              transaction.gas,
+              transaction.gas_price,
+              transaction.gas_used,
+              transaction.index,
+              transaction.input,
+              transaction.nonce,
+              transaction.r,
+              transaction.s,
+              transaction.status,
+              transaction.to_address_hash,
+              transaction.v,
+              transaction.value,
+              transaction.earliest_processing_start,
+              transaction.revert_reason,
+              transaction.max_priority_fee_per_gas,
+              transaction.max_fee_per_gas,
+              transaction.type,
+              transaction.l1_fee,
+              transaction.l1_fee_scalar,
+              transaction.l1_gas_price,
+              transaction.l1_gas_used,
+              transaction.l1_tx_origin,
+              transaction.l1_block_number
+            )
+        )
+
+      _ ->
+        from(
+          transaction in Transaction,
+          update: [
+            set: [
+              block_hash: fragment("EXCLUDED.block_hash"),
+              old_block_hash: transaction.block_hash,
+              block_number: fragment("EXCLUDED.block_number"),
+              block_consensus: fragment("EXCLUDED.block_consensus"),
+              block_timestamp: fragment("EXCLUDED.block_timestamp"),
+              created_contract_address_hash: fragment("EXCLUDED.created_contract_address_hash"),
+              created_contract_code_indexed_at: fragment("EXCLUDED.created_contract_code_indexed_at"),
+              cumulative_gas_used: fragment("EXCLUDED.cumulative_gas_used"),
+              error: fragment("EXCLUDED.error"),
+              from_address_hash: fragment("EXCLUDED.from_address_hash"),
+              gas: fragment("EXCLUDED.gas"),
+              gas_price: fragment("EXCLUDED.gas_price"),
+              gas_used: fragment("EXCLUDED.gas_used"),
+              index: fragment("EXCLUDED.index"),
+              input: fragment("EXCLUDED.input"),
+              nonce: fragment("EXCLUDED.nonce"),
+              r: fragment("EXCLUDED.r"),
+              s: fragment("EXCLUDED.s"),
+              status: fragment("EXCLUDED.status"),
+              to_address_hash: fragment("EXCLUDED.to_address_hash"),
+              v: fragment("EXCLUDED.v"),
+              value: fragment("EXCLUDED.value"),
+              earliest_processing_start: fragment("EXCLUDED.earliest_processing_start"),
+              revert_reason: fragment("EXCLUDED.revert_reason"),
+              max_priority_fee_per_gas: fragment("EXCLUDED.max_priority_fee_per_gas"),
+              max_fee_per_gas: fragment("EXCLUDED.max_fee_per_gas"),
+              type: fragment("EXCLUDED.type"),
+              # Don't update `hash` as it is part of the primary key and used for the conflict target
+              inserted_at: fragment("LEAST(?, EXCLUDED.inserted_at)", transaction.inserted_at),
+              updated_at: fragment("GREATEST(?, EXCLUDED.updated_at)", transaction.updated_at)
+            ]
+          ],
+          where:
+            fragment(
+              "(EXCLUDED.block_hash, EXCLUDED.block_number, EXCLUDED.block_consensus, EXCLUDED.block_timestamp, EXCLUDED.created_contract_address_hash, EXCLUDED.created_contract_code_indexed_at, EXCLUDED.cumulative_gas_used, EXCLUDED.from_address_hash, EXCLUDED.gas, EXCLUDED.gas_price, EXCLUDED.gas_used, EXCLUDED.index, EXCLUDED.input, EXCLUDED.nonce, EXCLUDED.r, EXCLUDED.s, EXCLUDED.status, EXCLUDED.to_address_hash, EXCLUDED.v, EXCLUDED.value, EXCLUDED.earliest_processing_start, EXCLUDED.revert_reason, EXCLUDED.max_priority_fee_per_gas, EXCLUDED.max_fee_per_gas, EXCLUDED.type) IS DISTINCT FROM (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)",
+              transaction.block_hash,
+              transaction.block_number,
+              transaction.block_consensus,
+              transaction.block_timestamp,
+              transaction.created_contract_address_hash,
+              transaction.created_contract_code_indexed_at,
+              transaction.cumulative_gas_used,
+              transaction.from_address_hash,
+              transaction.gas,
+              transaction.gas_price,
+              transaction.gas_used,
+              transaction.index,
+              transaction.input,
+              transaction.nonce,
+              transaction.r,
+              transaction.s,
+              transaction.status,
+              transaction.to_address_hash,
+              transaction.v,
+              transaction.value,
+              transaction.earliest_processing_start,
+              transaction.revert_reason,
+              transaction.max_priority_fee_per_gas,
+              transaction.max_fee_per_gas,
+              transaction.type
+            )
+        )
     end
   end
