+++ apps/explorer/lib/explorer/chain/optimism/frame_sequence.ex
@@ -1,16 +1,17 @@
-defmodule Explorer.Chain.OptimismFrameSequence do
+defmodule Explorer.Chain.Optimism.FrameSequence do
   @moduledoc "Models a frame sequence for Optimism."
 
   use Explorer.Schema
 
-  alias Explorer.Chain.{Hash, OptimismTxnBatch}
+  alias Explorer.Chain.Hash
+  alias Explorer.Chain.Optimism.TxnBatch
 
   @required_attrs ~w(id l1_transaction_hashes l1_timestamp)a
 
   @type t :: %__MODULE__{
           l1_transaction_hashes: [Hash.t()],
           l1_timestamp: DateTime.t(),
-          transaction_batches: %Ecto.Association.NotLoaded{} | [OptimismTxnBatch.t()]
+          transaction_batches: %Ecto.Association.NotLoaded{} | [TxnBatch.t()]
         }
 
   @primary_key {:id, :integer, autogenerate: false}
@@ -18,7 +19,7 @@ defmodule Explorer.Chain.OptimismFrameSequence do
     field(:l1_transaction_hashes, {:array, Hash.Full})
     field(:l1_timestamp, :utc_datetime_usec)
 
-    has_many(:transaction_batches, OptimismTxnBatch, foreign_key: :frame_sequence_id)
+    has_many(:transaction_batches, TxnBatch, foreign_key: :frame_sequence_id)
 
     timestamps()
   end
