+++ apps/indexer/lib/indexer/supervisor.ex
@@ -33,11 +33,6 @@ defmodule Indexer.Supervisor do
     EmptyBlocksSanitizer,
     InternalTransaction,
     Optimism,
-    OptimismDeposit,
-    OptimismOutputRoot,
-    OptimismTxnBatch,
-    OptimismWithdrawal,
-    OptimismWithdrawalEvent,
     PendingBlockOperationsSanitizer,
     PendingTransaction,
     ReplacedTransaction,
@@ -144,13 +139,13 @@ defmodule Indexer.Supervisor do
          [[json_rpc_named_arguments: json_rpc_named_arguments, memory_monitor: memory_monitor]]},
         {ReplacedTransaction.Supervisor, [[memory_monitor: memory_monitor]]},
         {Optimism.Supervisor, [[memory_monitor: memory_monitor]]},
-        {OptimismTxnBatch.Supervisor,
+        {Indexer.Fetcher.Optimism.TxnBatch.Supervisor,
          [[memory_monitor: memory_monitor, json_rpc_named_arguments: json_rpc_named_arguments]]},
-        {OptimismOutputRoot.Supervisor, [[memory_monitor: memory_monitor]]},
-        {OptimismDeposit.Supervisor, [[memory_monitor: memory_monitor]]},
-        {OptimismWithdrawal.Supervisor,
+        {Indexer.Fetcher.Optimism.OutputRoot.Supervisor, [[memory_monitor: memory_monitor]]},
+        {Indexer.Fetcher.Optimism.Deposit.Supervisor, [[memory_monitor: memory_monitor]]},
+        {Indexer.Fetcher.Optimism.Withdrawal.Supervisor,
          [[memory_monitor: memory_monitor, json_rpc_named_arguments: json_rpc_named_arguments]]},
-        {OptimismWithdrawalEvent.Supervisor, [[memory_monitor: memory_monitor]]},
+        {Indexer.Fetcher.Optimism.WithdrawalEvent.Supervisor, [[memory_monitor: memory_monitor]]},
         configure(PolygonEdge.Supervisor, [[memory_monitor: memory_monitor]]),
         {Indexer.Fetcher.RollupL1ReorgMonitor.Supervisor, [[memory_monitor: memory_monitor]]},
         configure(Indexer.Fetcher.PolygonEdge.Deposit.Supervisor, [[memory_monitor: memory_monitor]]),
