+++ config/runtime.exs
@@ -685,22 +685,22 @@ config :indexer, Indexer.Fetcher.Optimism,
   optimism_l1_rpc: System.get_env("INDEXER_OPTIMISM_L1_RPC"),
   optimism_l1_portal: System.get_env("INDEXER_OPTIMISM_L1_PORTAL_CONTRACT")
 
-config :indexer, Indexer.Fetcher.OptimismDeposit,
+config :indexer, Indexer.Fetcher.Optimism.Deposit,
   start_block_l1: System.get_env("INDEXER_OPTIMISM_L1_DEPOSITS_START_BLOCK"),
   batch_size: System.get_env("INDEXER_OPTIMISM_L1_DEPOSITS_BATCH_SIZE")
 
-config :indexer, Indexer.Fetcher.OptimismOutputRoot,
+config :indexer, Indexer.Fetcher.Optimism.OutputRoot,
   start_block_l1: System.get_env("INDEXER_OPTIMISM_L1_OUTPUT_ROOTS_START_BLOCK"),
   output_oracle: System.get_env("INDEXER_OPTIMISM_L1_OUTPUT_ORACLE_CONTRACT")
 
-config :indexer, Indexer.Fetcher.OptimismWithdrawal,
+config :indexer, Indexer.Fetcher.Optimism.Withdrawal,
   start_block_l2: System.get_env("INDEXER_OPTIMISM_L2_WITHDRAWALS_START_BLOCK"),
   message_passer: System.get_env("INDEXER_OPTIMISM_L2_MESSAGE_PASSER_CONTRACT")
 
-config :indexer, Indexer.Fetcher.OptimismWithdrawalEvent,
+config :indexer, Indexer.Fetcher.Optimism.WithdrawalEvent,
   start_block_l1: System.get_env("INDEXER_OPTIMISM_L1_WITHDRAWALS_START_BLOCK")
 
-config :indexer, Indexer.Fetcher.OptimismTxnBatch,
+config :indexer, Indexer.Fetcher.Optimism.TxnBatch,
   start_block_l1: System.get_env("INDEXER_OPTIMISM_L1_BATCH_START_BLOCK"),
   batch_inbox: System.get_env("INDEXER_OPTIMISM_L1_BATCH_INBOX"),
   batch_submitter: System.get_env("INDEXER_OPTIMISM_L1_BATCH_SUBMITTER"),
